+++ engine/src/bin/dsh_services.rs
@@ -4,6 +4,6 @@ use trifonius_engine::processor::ProcessorType;
 #[tokio::main]
 async fn main() {
   let processor_registry = ProcessorRegistry::default();
-  let application_descriptors = processor_registry.processor_descriptors_by_type(ProcessorType::Application);
-  println!("{}", serde_json::to_string_pretty(&application_descriptors).unwrap());
+  let dsh_service_descriptors = processor_registry.processor_descriptors_by_type(ProcessorType::DshService);
+  println!("{}", serde_json::to_string_pretty(&dsh_service_descriptors).unwrap());
 }