+++ engine/src/bin/status.rs
@@ -6,8 +6,8 @@ const SERVICE_ID: &str = "test-0-0-2";
 #[tokio::main]
 async fn main() -> Result<(), String> {
   let processor_registry = ProcessorRegistry::default();
-  let application = processor_registry.processor(ProcessorType::Application, "consentfilter").ok_or("")?;
-  let status = application.status(SERVICE_ID).await?;
+  let dsh_service = processor_registry.processor(ProcessorType::DshService, "consentfilter").ok_or("")?;
+  let status = dsh_service.status(SERVICE_ID).await?;
   println!("{}", status);
   Ok(())
 }