+++ engine/src/resource/dsh_topic/topic_registry.rs
@@ -2,7 +2,7 @@ use std::collections::HashMap;
 
 use dsh_sdk::Properties;
 
-use crate::processor::application::TargetClientFactory;
+use crate::processor::dsh_service::TargetClientFactory;
 use crate::resource::dsh_topic::topic_resource::{topic_resource_identifier, TopicResourceImpl};
 use crate::resource::resource::{Resource, ResourceIdentifier, ResourceStatus};
 use crate::resource::resource_descriptor::ResourceDescriptor;